@@ -1,4 +1,56 @@
+[build-system]
+requires = ["poetry-core>=1.0.0"]
+build-backend = "poetry.core.masonry.api"
+
 [tool.black]
 line-length = 79
-skip-string-normalization = true
-target-version = ["py36", "py37", "py38"]
+target-version = ["py38", "py39"]
+
+[tool.isort]
+ensure_newline_before_comments = true
+line_length = 79
+multi_line_output = 3
+include_trailing_comma = true
+skip_glob = ["docs/*", "migrations/*", "*.egg-info"]
+
+[tool.poetry]
+name = "poc-django"
+version = "0.1.0"
+description = ""
+authors = ["Ivan Ogasawara <ivan.ogasawara@gmail.com>"]
+license = "MIT"
+
+[tool.poetry.dependencies]
+python = "^3.9"
+Django = "^4.*"
+psycopg2-binary = "^2.8.6"
+pytz = "^2021.1"
+PyYAML = "^5.4.1"
+redis = "^4"
+regex = "^2021.4.4"
+requests = "^2.25.1"
+requests-oauthlib = "^1.3.0"
+toml = "^0.10.2"
+djangorestframework = "^3.13.1"
+django-cors-headers = "^3.11.0"
+
+[tool.poetry.dev-dependencies]
+black = ">=20.8b1"
+coverage = "^6.3.1"
+django-extensions = "^3.1.3"
+django-debug-toolbar = "^3.2.1"
+flake8 = ">=3.9.1"
+isort = ">=5.8.0"
+mypy = ">=0.812"
+mypy-extensions = "^0.4.3"
+pre-commit = "^2.12.1"
+pycodestyle = "^2.7.0"
+pytest = ">=6.2.3"
+pytest-django = ">=4.2.0"
+pytest-sugar = ">=0.9.4"
+typing-extensions = "^4"
+Werkzeug = "^2"
+whitenoise = ">5.2.0"
+pytest-cov = "^3.0.0"
+factory-boy = "^3"
+django-upgrade = "^1.5.0"
