@@ -1,16 +1,40 @@
 """
-ASGI config for djangopoc project.
+ASGI config for PoC Django project.
 
 It exposes the ASGI callable as a module-level variable named ``application``.
 
 For more information on this file, see
-https://docs.djangoproject.com/en/3.0/howto/deployment/asgi/
-"""
+https://docs.djangoproject.com/en/dev/howto/deployment/asgi/
 
+"""
 import os
+import sys
+from pathlib import Path
 
 from django.core.asgi import get_asgi_application
 
-os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'config.settings')
+# This allows easy placement of apps within the interior
+# poc_django directory.
+ROOT_DIR = Path(__file__).resolve(strict=True).parent.parent
+sys.path.append(str(ROOT_DIR / "poc_django"))
 
-application = get_asgi_application()
+# If DJANGO_SETTINGS_MODULE is unset, default to the local settings
+os.environ.setdefault("DJANGO_SETTINGS_MODULE", "config.settings.dev")
+
+# This application object is used by any ASGI server configured to use this file.
+django_application = get_asgi_application()
+# Apply ASGI middleware here.
+# from helloworld.asgi import HelloWorldApplication
+# application = HelloWorldApplication(application)
+
+# Import websocket application here, so apps from django_application are loaded first
+from config.websocket import websocket_application  # noqa isort:skip
+
+
+async def application(scope, receive, send):
+    if scope["type"] == "http":
+        await django_application(scope, receive, send)
+    elif scope["type"] == "websocket":
+        await websocket_application(scope, receive, send)
+    else:
+        raise NotImplementedError(f"Unknown scope type {scope['type']}")
